Add flag parsing
Add a bunch of flags for opting to *not* use specific sensors, and
change the Prometheus port. This hasn't been thoroughly tested, but
seems to at least not regress in the "I want both" case wit no Python
metrics exposed.

Add an environment file and use its ARGS variable in the systemd
service.

Add a restart directive for the service as well, since that was missing
and the default is not to.

Update the README now that configuration isn't "hack about the code how
you like it".