Persist the SGP30 baseline
Set up a /var/lib/ directory, and use it to JSON-persist the baseline
values from the SGP30, reading at startup, and saving every hour (and at
startup). Includes checking a timestamp (mtime is not trusted---we care
about when the reading was taken, not the file touched) to ignore values
that are too old, and any other file-related crimes.

It is not (currently) exported to Prometheus, since it's updated much
less frequently, and in theory just boring internal details.