Perform humidity compensation for the SGP30
The driver integration guide is pretty helpful for this one, the formula
seems to work right. Calculate the absolute humidity and feed it to the
SGP30 to compensate for. If out of range, disable compensation and warn
(although stderr to syslog still seems broken, annoyingly). It's traced
when using --output-stdout (mostly for debugging/verification), but not
exported to Prometheus, mainly because it's a derived value from others
that are being exported, so could just be calculated at
graphing/alerting time rather than stored, if you cared. (Relative
humidity is generally much more useful for how the air feels.)