Move console graphs to flexbox
This seems to behave a lot better in Mobile Firefox, and lines up the
things I care about nicely on desktop too. The smaller size of the
graphs doesn't matter tremendously since you can always click through to
the ad-hoc query graphs for full width display.

Update the screenshot to match. Add a line that it's customizable.