Display tweaks that I'm mostly happy with
Increase the timing before rotating to the next metric. I was a little
suspicious that the delay wasn't not applying properly; this didn't
feel like three seconds. Python's time.sleep() is supposed to resume
after signals if needed. Then I got another timesource to compare and
found that it's just my brain being underclocked, apparently. It works.

Adjust the default ratio of metric-to-units to 3:1.

Add an experimental reverse-video bottom bar, but I don't really like
it.

Blank the display on exit so it doesn't show lies if the Pi shuts down.
It seems to power the backlight, but not the GPIO pin that tells the
backlight to stay off.