More small graphical tweaks
Use thicker arrows, which are present in DejaVuSans, although weirdly
the right-pointing version is in a non-contiguous codepoint in Unicode,
and a missing glyph; so leave that less bold.

Experimented with smallcaps PPM, but not the best effect.

Measure lower-bar font size more accurately by hardcoding which is the
longest unit.

Further rebalance the bottom bar size for good defaults.