Split top and bottom fonts
This allows using a fancier font with less Unicode coverage for the main
digit display, while still using fancy arrow glyphs or such for the
trend indicators.

Try using PIL's textbbox() to avoid ascender spacing, but it's not
implemented in the version in the python3-pil package in Debian buster,
it seems.