Add systemd conf for trend display; fix logging
Add an environment file and systemd unit for the trend display. I'm
aggressive about namespacing, so it gets to be huge. You can always
rename it when you install it.

Update the instructions to suggest using it.

Fix the environment files having DOS line endings for some reason.

Fix logging. It turns out the problem is good old standard I/O stream
buffering...I'm not sure why Python *never* flushes when talking to
journald, but unbuffering those streams fixes it.