Use LTR599 to turn off ST7789 at night
If it's dark enough, turn off the backlight.
If something is near the screen (like a hand waving at it), turn it on
temporarily.

Good initial values seem to be --sleep-lux 3 --wake-proximity 0.01

Add support for displaying SCD4X metrics. This is currently a bit gross
and copy-paste-y.

Fix exporting the SCD4X metrics under BME280 names (oops).

Update copyright years on touched files.