Add a cputemp compensation patch.
Yes this could be a branch, but it's an expected dead-end.

In practice, this just doesn't really do the right thing. There's a
consistent misread from the SCD41, but it's more a linear offset than a
CPU-dependent one. Tuning it for CPU idle caused bad sagging for CPU
rampup. I suspect much of the heat is actually from the other components
on the breakout garden.