Allow setting the SCD4X temperature offset
An adjustment of 7.5 seems to work for me for the breakout garden with a
ST7789, LTR559, atop a Pi Zero W seperated by the case top and a boost
header.

This is slightly tricky to set; there is a default offset of 4 to allow
for. See the datasheet. Do not save the value to EEPROM. Read whatever
*was* set and log it.

Do the same for altitude, since it exists.